Change CONTENT_TAGS to blacklist
Previously, if you use the KEEP_CONTENT option, the tag had to match a
whitelist for its contents to be kept. This is not necessary for security,
since the contents is still sanitised just like any other content in the page,
and is annoying if you are sanitising random HTML from sources which may use
arbitrary non-standard tags to wrap bits of real content (resulting in chunks
of text being stripped by the sanitiser).

This commit changes the CONTENT_TAGS whitelist to a FORBID_CONTENTS blacklist;
the blacklist is not needed for security, but we do want to remove the contents
of <script> tags, style tags (if <style> has been disallowed) and anything in
the <head> (if disallowed), since we know these are not designed to have
user-visible content.

Implements #65.