Fix "forbidden" attributes allowed if ALLOW_UNKNOWN_PROTOCOLS true
The logic for whether to forbid an attribute was broken: if
ALLOW_UNKNOWN_PROTOCOLS was true, then it would allow any attribute with
a value that didn't match the IS_SCRIPT_OR_DATA regexp.

This evades the attribute name whitelist, and is a security flaw: for
example you could set an "onX" JavaScript attribute.

The logic here got too convoluted to be easy to read, which is
presumably how this bug crept in. I have rewritten it into an expanded
set of if/else statements which I believe are easier to follow.